requirements.txt
google-cloud-storage==2.14.0
flask==2.3.3
functions-framework==3.5.0
gunicorn==21.2.0
